+++ src/z3c/autoinclude/utils.py
@@ -39,7 +39,7 @@ class ZCMLInfo(dict):
         for zcml_group in zcml_to_look_for:
             self[zcml_group] = []
 
-    
+
 def subpackageDottedNames(package_path, ns_dottedname=None):
     # we do not look for subpackages in zipped eggs
     if not isUnzippedEgg(package_path):
@@ -79,17 +79,24 @@ def distributionForDottedName(package_dottedname):
         for dist in dists:
             if not isUnzippedEgg(dist.location):
                 continue
-            packages = find_packages(dist.location)
             ns_packages = namespaceDottedNames(dist)
+            packages = find_packages(dist.location)
             #if package_dottedname in ns_packages:
                 #continue
             if package_dottedname not in packages:
                 continue
+            if dist.key.lower() != package_dottedname.lower():
+                if not dist.key.lower().startswith(package_dottedname.lower()):
+                    continue
+                # make sure that the dottedname bla.blubb
+                # matches bla.blubb.xx but not bla.blubber
+                if dist.key[len(package_dottedname)] != '.':
+                    continue
             valid_dists_for_package.append((dist, ns_packages))
 
     if len(valid_dists_for_package) == 0:
         raise LookupError("No distributions found for package `%s`; are you sure it is importable?" % package_dottedname)
-    
+
     if len(valid_dists_for_package) > 1:
         non_namespaced_dists = filter(lambda x: len(x[1]) is 0, valid_dists_for_package)
         if len(non_namespaced_dists) == 0:
@@ -126,7 +133,7 @@ def namespaceDottedNames(dist):
     except KeyError:
         ns_dottednames = []
     return ns_dottednames
-    
+
 def isUnzippedEgg(path):
     """
     Check whether a filesystem path points to an unzipped egg; z3c.autoinclude
@@ -136,6 +143,7 @@ def isUnzippedEgg(path):
     """
     return os.path.isdir(path)
 
+CACHE = {}
 ### cargo-culted from setuptools 0.6c9's __init__.py;
 #   importing setuptools is unsafe, but i can't find any
 #   way to get the information that find_packages provides
@@ -151,17 +159,22 @@ def find_packages(where='.', exclude=()):
     names, such that 'foo.*' will exclude all subpackages of 'foo' (but not
     'foo' itself).
     """
+    original_where = where
+    if where in CACHE:
+        return CACHE[where]
     out = []
     stack=[(convert_path(where), '')]
     while stack:
         where,prefix = stack.pop(0)
+        os.listdir(where)
         for name in os.listdir(where):
             fn = os.path.join(where,name)
-            if ('.' not in name and os.path.isdir(fn) and
-                os.path.isfile(os.path.join(fn,'__init__.py'))
-            ):
-                out.append(prefix+name); stack.append((fn,prefix+name+'.'))
+            if ('.' not in name and os.path.isdir(fn)):
+                stack.append((fn, prefix+name+'.'))
+                if os.path.isfile(os.path.join(fn, '__init__.py')):
+                    out.append(prefix+name)
     for pat in list(exclude)+['ez_setup']:
         from fnmatch import fnmatchcase
         out = [item for item in out if not fnmatchcase(item,pat)]
+    CACHE[original_where] = out
     return out